pyproject.toml
[tool.poetry]
name = "django-drf"
version = "0.1.0"
description = ""
authors = ["Clayton Silva"]
license = "MIT"
readme = "README.md"
packages = [{include = "django_drf"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.3"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
drf-spectacular = {extras = ["sidecar"], version = "^0.24.2"}
drf-yasg = "^1.21.4"
swagger-spec-validator = "^3.0.3"
drf-spectacular-sidecar = "^2022.11.1"
mysqlclient = "^2.1.1"
djangorestframework-jwt = "^1.11.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
black = "^22.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
